requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.10.1
django-jazzmin==2.4.8
django-phonenumber-field==6.0.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
phonenumbers==8.12.40
Pillow==8.4.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
